3. django/1026/code/requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.12.4
Faker==9.5.2
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.21
psycopg2==2.9.1
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
traitlets==5.1.0
wcwidth==0.2.5
